fix double ostacoli

+++ src/colorclash/model/Particle.java
@@ -36,12 +36,12 @@ public class Particle {
         }
     }
 
-    public int getX() {
-        return (int) x;
+    public double getX() {
+        return  x;
     }
 
-    public int getY() {
-        return (int) y;
+    public double getY() {
+        return  y;
     }
 
     public int getSize() {
